test: split and clean up
